requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers==4.21.3
pandas==1.4.4
numpy==1.23.3
scikit-learn==1.1.2
Pillow==9.2.0
matplotlib==3.5.3
tqdm==4.64.1
openpyxl==3.0.10
seaborn>=0.11.2 